# CI Pipelines build with WDK Nuget Packages