With the WDK now supporting Nuget packages for ease of install it is beneficial to have the CI pipeline also utilize this ability. This will allow easier and faster integration when new WDK versions are published without relying on the runner images to be updated.

- Adds Nuget package capability to the CI pipeline.
- Adds installation of WDK vsix directly into the CI pipeline.
- Updates environment variables for ease of incrementing component versions.

### Pipeline Time Metrics
**BuildAll** (push and pull_request)
Without Vsix Install: ~08m 59s
With Vsix Install: ~38m 59s (optimized ~8m)

**CodeQL** (push and pull_request)
Without Vsix Install: ~1h 1m 43s
With Vsix Install: ~1h 34m 31s (optimized ~1h 16m 2s)

**Average** WDK Vsix Install Time
~21m (optimized ~15s)

**Average** WDK Nuget Restore
~1m 30s